iot-azure-hdt/requirements.txt
azure-digitaltwins-core
azure-identity
pyserial
web3
cryptography
dotenv